test(backend): move TestClockProvider out
